pyproject.toml
[project]
name = "brainstorming-ma"
version = "0.1.0"
description = "Multi-Agent Brainstorming System using A2A Protocol"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "google-generativeai>=0.3.0",
    "streamlit>=1.28.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.25.0",
]
